[FEAT] Puzzle2

+++ 4Q_Solution/Client/BaseWoolball.cpp
@@ -64,7 +64,7 @@ void BaseWoolball::InteractCallback(const PlayMsg::InteractObject* msg)
 	if (!isData)
 		return;
 	auto data = std::any_cast<GameCoreData>(*isData);
-	if (data.player == PlayerEnum::Live || data.player == PlayerEnum::Developer)
+	if (data.player == static_cast<int>(PlayerEnum::Live) || data.player == static_cast<int>(PlayerEnum::Developer))
 	{
 		if (_posIndex != muteValue[_index])
 		{
